LayoutTests/fast/css/z-index-calculated-value-expected.txt

Tests assigning a calculated value to z-index CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['z-index'] is ""
testDiv.style['z-index'] = 'calc(-2 * 3)'
PASS testDiv.style['z-index'] is "-6"
PASS window.getComputedStyle(testDiv).getPropertyValue('z-index') is "-6"
PASS successfullyParsed is true

TEST COMPLETE
